wip: add utxo confirmed endpoint

+++ modules/fedimint-wallet-server/src/lib.rs
@@ -77,7 +77,7 @@ use fedimint_wallet_common::config::{WalletClientConfig, WalletConfig, WalletGen
 use fedimint_wallet_common::endpoint_constants::{
     ACTIVATE_CONSENSUS_VERSION_VOTING_ENDPOINT, BITCOIN_KIND_ENDPOINT, BITCOIN_RPC_CONFIG_ENDPOINT,
     BLOCK_COUNT_ENDPOINT, BLOCK_COUNT_LOCAL_ENDPOINT, MODULE_CONSENSUS_VERSION_ENDPOINT,
-    PEG_OUT_FEES_ENDPOINT, WALLET_SUMMARY_ENDPOINT,
+    PEG_OUT_FEES_ENDPOINT, UTXO_CONFIRMED_ENDPOINT, WALLET_SUMMARY_ENDPOINT,
 };
 use fedimint_wallet_common::keys::CompressedPublicKey;
 use fedimint_wallet_common::tweakable::Tweakable;
@@ -270,7 +270,8 @@ impl ServerModuleInit for WalletInit {
                 MODULE_CONSENSUS_VERSION.major,
                 MODULE_CONSENSUS_VERSION.minor,
             ),
-            &[(0, 1)],
+            // TODO: separate commit that bumps this version
+            &[(0, 2)],
         )
     }
 
@@ -822,29 +823,6 @@ impl ServerModule for Wallet {
     fn api_endpoints(&self) -> Vec<ApiEndpoint<Self>> {
         vec![
             api_endpoint! {
-                MODULE_CONSENSUS_VERSION_ENDPOINT,
-                ApiVersion::new(0, 0),
-                async |module: &Wallet, context, _params: ()| -> ModuleConsensusVersion {
-                    Ok(module.consensus_module_consensus_version(&mut context.dbtx().into_nc()).await)
-                }
-            },
-            api_endpoint! {
-                ACTIVATE_CONSENSUS_VERSION_VOTING_ENDPOINT,
-                ApiVersion::new(0, 0),
-                async |_module: &Wallet, context, _params: ()| -> () {
-                    check_auth(context)?;
-
-                    let mut dbtx = context.dbtx();
-
-                    dbtx.insert_entry(&ConsensusVersionVotingActivationKey, &()).await;
-
-                    dbtx.commit_tx_result().await.map_err(|e| ApiError::server_error(e.to_string()))?;
-
-                    Ok(())
-
-                }
-            },
-            api_endpoint! {
                 BLOCK_COUNT_ENDPOINT,
                 ApiVersion::new(0, 0),
                 async |module: &Wallet, context, _params: ()| -> u32 {
@@ -921,6 +899,36 @@ impl ServerModule for Wallet {
                 ApiVersion::new(0, 1),
                 async |module: &Wallet, context, _params: ()| -> WalletSummary {
                     Ok(module.get_wallet_summary(&mut context.dbtx().into_nc()).await)
+                }
+            },
+            api_endpoint! {
+                MODULE_CONSENSUS_VERSION_ENDPOINT,
+                ApiVersion::new(0, 2),
+                async |module: &Wallet, context, _params: ()| -> ModuleConsensusVersion {
+                    Ok(module.consensus_module_consensus_version(&mut context.dbtx().into_nc()).await)
+                }
+            },
+            api_endpoint! {
+                ACTIVATE_CONSENSUS_VERSION_VOTING_ENDPOINT,
+                ApiVersion::new(0, 2),
+                async |_module: &Wallet, context, _params: ()| -> () {
+                    check_auth(context)?;
+
+                    let mut dbtx = context.dbtx();
+
+                    dbtx.insert_entry(&ConsensusVersionVotingActivationKey, &()).await;
+
+                    dbtx.commit_tx_result().await.map_err(|e| ApiError::server_error(e.to_string()))?;
+
+                    Ok(())
+
+                }
+            },
+            api_endpoint! {
+                UTXO_CONFIRMED_ENDPOINT,
+                ApiVersion::new(0, 2),
+                async |module: &Wallet, context, outpoint: bitcoin::OutPoint| -> bool {
+                    Ok(module.is_utxo_confirmed(&mut context.dbtx().into_nc(), outpoint).await)
                 }
             },
         ]
@@ -1246,34 +1254,41 @@ impl Wallet {
                 .await
                 .expect("bitcoind rpc to get block hash");
 
-            let block = retry("get_block", backoff_util::background_backoff(), || {
-                self.btc_rpc.get_block(&block_hash)
-            })
-            .await
-            .expect("bitcoind rpc to get block");
+            if self.consensus_module_consensus_version(dbtx).await
+                >= ModuleConsensusVersion::new(2, 2)
+            {
+                let block = retry("get_block", backoff_util::background_backoff(), || {
+                    self.btc_rpc.get_block(&block_hash)
+                })
+                .await
+                .expect("bitcoind rpc to get block");
 
-            for transaction in block.txdata {
-                // We maintain the subset of unspent P2WSH transaction outputs created
-                // since the federation was established in the database.
+                for transaction in block.txdata {
+                    // We maintain the subset of unspent P2WSH transaction outputs created
+                    // since the federation was established in the database.
 
-                for tx_in in &transaction.input {
-                    dbtx.remove_entry(&UnspentTxOutKey(tx_in.previous_output))
-                        .await;
-                }
+                    for tx_in in &transaction.input {
+                        dbtx.remove_entry(&UnspentTxOutKey(tx_in.previous_output))
+                            .await;
+                    }
 
-                for (vout, tx_out) in transaction.output.iter().enumerate() {
-                    if if self.cfg.consensus.peer_peg_in_keys.len() > 1 {
-                        tx_out.script_pubkey.is_p2wsh()
-                    } else {
-                        tx_out.script_pubkey.is_p2wpkh()
-                    } {
-                        let outpoint = bitcoin::OutPoint {
-                            txid: transaction.compute_txid(),
-                            vout: vout as u32,
+                    for (vout, tx_out) in transaction.output.iter().enumerate() {
+                        // TODO: separate commit for refactor
+                        let should_track_utxo = if self.cfg.consensus.peer_peg_in_keys.len() > 1 {
+                            tx_out.script_pubkey.is_p2wsh()
+                        } else {
+                            tx_out.script_pubkey.is_p2wpkh()
                         };
 
-                        dbtx.insert_new_entry(&UnspentTxOutKey(outpoint), tx_out)
-                            .await;
+                        if should_track_utxo {
+                            let outpoint = bitcoin::OutPoint {
+                                txid: transaction.compute_txid(),
+                                vout: vout as u32,
+                            };
+
+                            dbtx.insert_new_entry(&UnspentTxOutKey(outpoint), tx_out)
+                                .await;
+                        }
                     }
                 }
             }
@@ -1532,6 +1547,15 @@ impl Wallet {
             unconfirmed_peg_out_txos,
             unconfirmed_change_utxos,
         }
+    }
+
+    // TODO: how to test better?
+    async fn is_utxo_confirmed(
+        &self,
+        dbtx: &mut DatabaseTransaction<'_>,
+        outpoint: bitcoin::OutPoint,
+    ) -> bool {
+        dbtx.get_value(&UnspentTxOutKey(outpoint)).await.is_some()
     }
 
     fn offline_wallet(&self) -> StatelessWallet {
